+++ lib/modules/asset/types/AssetApi.ts
@@ -1,6 +1,6 @@
 import { JSONObject, KDocument, KHit, SearchResult } from "kuzzle-sdk";
 
-import { MeasureContent } from "../../../modules/measure";
+import { MeasureContent, Measurement } from "../../../modules/measure";
 import {
   ApiDigitalTwinGetLastMeasuredAtRequest,
   ApiDigitalTwinGetLastMeasuredAtResult,
@@ -16,11 +16,18 @@ import {
 import { AssetContent } from "./AssetContent";
 
 type AssetsControllerName = "device-manager/assets";
+import { ApiMeasureSource } from "../../measure/types/MeasureSources";
 
 interface AssetsControllerRequest {
   controller: AssetsControllerName;
 
   engineId: string;
+
+  /**
+   * ? Request parameter used by SoftTenant module
+   * ! Not used directly in this plugin
+   */
+  softTenantId?: string | string[] | null;
 }
 
 export interface ApiAssetGetRequest extends AssetsControllerRequest {
@@ -59,8 +66,6 @@ export type ApiAssetMetadataReplaceResult = KDocument<AssetContent>;
 
 export interface ApiAssetUpsertRequest extends AssetsControllerRequest {
   action: "upsert";
-
-  _id: string;
 
   refresh?: string;
 
@@ -141,6 +146,43 @@ export type ApiAssetGetMeasuresResult = {
   measures: Array<KDocument<MeasureContent<JSONObject>>>;
   total: number;
 };
+
+type TypelessApiMeasureSource = Omit<ApiMeasureSource, "type">;
+
+export interface ApiAssetMeasureIngestRequest extends AssetsControllerRequest {
+  action: "measureIngest";
+
+  assetId: string;
+
+  engineId: string;
+  engineGroup?: string;
+  slotName: string;
+
+  body: {
+    dataSource: TypelessApiMeasureSource;
+    measuredAt: number;
+    values: JSONObject;
+  };
+}
+export type ApiAssetMeasureIngestResult = void;
+
+type APIDecodedMeasurement = Omit<Measurement, "type"> & { slotName: string };
+
+export interface ApiAssetmMeasureIngestRequest extends AssetsControllerRequest {
+  action: "mMeasureIngest";
+
+  assetId: string;
+
+  engineId: string;
+  engineGroup?: string;
+
+  body: {
+    dataSource: TypelessApiMeasureSource;
+    measurements: APIDecodedMeasurement[];
+  };
+}
+
+export type ApiAssetmMeasureIngestResult = void;
 
 export type ApiAssetGetLastMeasuresRequest =
   ApiDigitalTwinGetLastMeasuresRequest<AssetsControllerName>;