+++ tests/application/measures/Magicule.ts
@@ -1,0 +1,22 @@
+import { MeasureModel } from "../../../index";
+
+export type MagiculeMeasurement = {
+  magicule: number;
+};
+
+export const Magicule: MeasureModel = {
+  modelName: "magicule",
+  definition: {
+    valuesMappings: {
+      magicule: { type: "integer" },
+    },
+    validationSchema: {
+      type: "object",
+      properties: {
+        magicule: { type: "integer" },
+      },
+      required: ["magicule"],
+      additionalProperties: false,
+    },
+  },
+};