+++ tests/scenario/modules/models/asset-model-metadata-propagation.test.ts
@@ -11,7 +11,7 @@ import {
 } from "../../../../lib/modules/model";
 import { setupHooks } from "../../../helpers";
 
-jest.setTimeout(10000);
+jest.setTimeout(20000);
 
 describe("Asset model metadata propagation", () => {
   const sdk = setupHooks();